6.21.7 Fix code review comments - Перенес OffersCardsListWithActive в экспорт компонента OffersCardsList.
