+++ src/components/offers-cards-list/offers-cards-list.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import {connect} from 'react-redux';
 
 import OffersCardsContainer from '../offers-cards-container/offers-cards-container.jsx';
 import OffersCardsListEmpty from '../offers-cards-list-empty/offers-cards-list-empty.jsx';
 import Map from '../map/map.jsx';
-import {withClassName} from '../../hocs/with-class-name/with-class-name.jsx';
 import SortOffersMenu from '../sort-offers-menu/sort-offers-menu.jsx';
+import {withClassName} from '../../hocs/with-class-name/with-class-name.jsx';
+import withActiveItem from '../../hocs/with-active-item/with-active-item.jsx';
+import {ActionCreator} from "../../reducer.js";
+import {SORT_TYPE} from '../../data/constants.js';
 
 const OffersCardsList = (props) => {
   const {sortType, items, activeItem, onOfferTitleClick, onSort, onItemSelected} = props;
@@ -77,4 +81,32 @@ OffersCardsList.propTypes = {
   ),
 };
 
-export default OffersCardsList;
+const sortOffers = (offers, sortType) => {
+  const copyOffers = [...offers];
+
+  switch (sortType) {
+    case SORT_TYPE.PRICE_HIGH_TO_LOW:
+      return copyOffers.sort((a, b) => b.price - a.price);
+    case SORT_TYPE.PRICE_LOW_TO_HIGH:
+      return copyOffers.sort((a, b) => a.price - b.price);
+    case SORT_TYPE.TOP_RATED:
+      return copyOffers.sort((a, b) => b.rating - a.rating);
+  }
+
+  return offers;
+};
+
+const mapStateToProps = (state) => ({
+  items: sortOffers(state.filteredOffers, state.sortType),
+  sortType: state.sortType
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  onSort(sortType) {
+    dispatch(ActionCreator.changeSortType(sortType));
+  }
+});
+
+export {OffersCardsList};
+
+export default connect(mapStateToProps, mapDispatchToProps)(withActiveItem(OffersCardsList));
