+++ src/components/offers-cards-list/offers-cards-list.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import OffersCardsList from './offers-cards-list.jsx';
+import {OffersCardsList} from './offers-cards-list.jsx';
 import {OFFERS_TESTS} from '../../mocks/offers-tests';
 import {SORT_TYPE} from '../../data/constants';
